Use mdx util package for sharing utils
