+++ packages/mdx/package.json
@@ -49,6 +49,7 @@
     "@babel/plugin-transform-react-jsx": "7.3.0",
     "@mapbox/rehype-prism": "0.3.1",
     "@mdx-js/react": "^1.1.6",
+    "@mdx-js/util": "^1.1.6",
     "babel-plugin-apply-mdx-type-prop": "^1.1.6",
     "babel-plugin-extract-import-names": "^1.1.6",
     "babel-plugin-remove-export-keywords": "^1.1.6",
