--- packages/mdx/test/util.test.js
@@ -1,14 +1,0 @@
-const {paramCase} = require('../util')
-
-const PARAM_CASE_FIXTURES = [
-  ['ariaHidden', 'aria-hidden'],
-  ['data123', 'data-123']
-]
-
-PARAM_CASE_FIXTURES.forEach(([input, expected]) => {
-  test(`it turns ${input} into ${expected}`, () => {
-    const result = paramCase(input)
-
-    expect(result).toEqual(expected)
-  })
-})
