+++ packages/remark-mdx/package.json
@@ -32,6 +32,7 @@
     "@babel/helper-plugin-utils": "7.0.0",
     "@babel/plugin-proposal-object-rest-spread": "7.5.5",
     "@babel/plugin-syntax-jsx": "7.2.0",
+    "@mdx-js/util": "^1.1.6",
     "is-alphabetical": "1.0.3",
     "remark-parse": "7.0.0",
     "unified": "8.3.2"
