+++ packages/util/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "babel-plugin-apply-mdx-type-prop",
+  "name": "@mdx-js/util",
   "version": "1.1.6",
-  "description": "Apply the MDX type prop used in the MDX pragma",
+  "description": "Utilities for MDX libraries",
   "license": "MIT",
   "keywords": [
     "mdx",
@@ -9,7 +9,7 @@
     "react",
     "jsx",
     "remark",
-    "babel"
+    "mdxast"
   ],
   "homepage": "https://mdxjs.com",
   "repository": "mdx-js/mdx",
@@ -19,13 +19,7 @@
     "index.js"
   ],
   "devDependencies": {
-    "@babel/core": "7.5.5",
-    "@babel/plugin-syntax-jsx": "^7.2.0",
     "jest": "24.8.0"
-  },
-  "dependencies": {
-    "@babel/helper-plugin-utils": "^7.0.0",
-    "@mdx-js/util": "^1.1.6"
   },
   "scripts": {
     "test": "jest"
